+++ examples/prisma/stacks/index.ts
@@ -3,7 +3,7 @@ import { MyStack } from "./MyStack";
 
 export default function main(app: App) {
   app.setDefaultFunctionProps({
-    runtime: "nodejs14.x",
+    runtime: "nodejs16.x",
     srcPath: "backend",
   });
   app.stack(MyStack);