+++ packages/create-sst/bin/presets/base/typescript/preset.mjs
@@ -13,7 +13,7 @@ export default [
   extend("presets/base/javascript"),
   extract(),
   install({
-    packages: ["typescript", "@tsconfig/node14"],
+    packages: ["typescript", "@tsconfig/node16"],
     dev: true,
   }),
   patch({