+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @serverless-stack/cli
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`b81522a4`](https://github.com/serverless-stack/serverless-stack/commit/b81522a4dd0789292a4300b5465c4cab13f7a0cc)]:
+  - @serverless-stack/core@1.1.1
+  - @serverless-stack/resources@1.1.1
+
+## 1.1.0
+
+### Minor Changes
+
+- 5860eb64: Update CDK to 2.24.0
+
+### Patch Changes
+
+- Updated dependencies [5860eb64]
+  - @serverless-stack/core@1.1.0
+  - @serverless-stack/resources@1.1.0
+
+## 1.0.12
+
+### Patch Changes
+
+- 31ec3eb1: Added release snapshot script
+- Updated dependencies [31ec3eb1]
+  - @serverless-stack/core@1.0.12
+  - @serverless-stack/resources@1.0.12
+
 ## 1.0.11
 
 ### Patch Changes