+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @serverless-stack/core
 
+## 1.1.1
+
+### Patch Changes
+
+- [#1698](https://github.com/serverless-stack/serverless-stack/pull/1698) [`b81522a4`](https://github.com/serverless-stack/serverless-stack/commit/b81522a4dd0789292a4300b5465c4cab13f7a0cc) Thanks [@thdxr](https://github.com/thdxr)! - My test feature
+
+## 1.1.0
+
+### Minor Changes
+
+- 5860eb64: Update CDK to 2.24.0
+
+## 1.0.12
+
+### Patch Changes
+
+- 31ec3eb1: Added release snapshot script
+
 ## 1.0.11
 
 ### Patch Changes