+++ packages/create-sst/CHANGELOG.md
@@ -1,3 +1,21 @@
 # create-sst
 
+## 1.1.1
+
+### Patch Changes
+
+- [`eb009dae`](https://github.com/serverless-stack/serverless-stack/commit/eb009dae1e182647aa9e76dacfff7d09a877b731) Thanks [@thdxr](https://github.com/thdxr)! - Update starters + examples to node16
+
+## 1.1.0
+
+### Minor Changes
+
+- 5860eb64: Update CDK to 2.24.0
+
+## 1.0.12
+
+### Patch Changes
+
+- 31ec3eb1: Added release snapshot script
+
 ## 1.0.11