+++ packages/static-site-env/CHANGELOG.md
@@ -1,3 +1,17 @@
 # @serverless-stack/static-site-env
 
+## 1.1.1
+
+## 1.1.0
+
+### Minor Changes
+
+- 5860eb64: Update CDK to 2.24.0
+
+## 1.0.12
+
+### Patch Changes
+
+- 31ec3eb1: Added release snapshot script
+
 ## 1.0.11